openai/requirements.txt
openai
openai-whisper
pytube